Let Int(\l)Column take a lambda with `const contact*`.
The second lambda with an extra argument is required at a
couple of places but we do not want to add the extra (unused)
argument everywhere an IntColumn is instantiated.

The current patch has the advantage of being transparent
to the callers.

A better solution would be to not require the `auth_user` but
there is currently no easy way to do this.

CMK-7076

Change-Id: Ie4ccc1f56eca31adbf3f94448956c431f2cc27ac