12129 FIX Fix discovered host labels file names
The discovered host labels of hosts that have a dot in their name were not
stored correctly. Instead of "abc.tld.mk" the labels were stored in a file
named "abc.mk". The name of these files have now been fixed. During update
of Checkmk the names of existing will be checked and updated.

Change-Id: I6738ad13e703715c0d0680f16a23724c13fc96cf