12167 FIX Fixed acknowledgement expiry
Previously, the expiration mechanism for acknowledgements ignored the
"persisten" flag, so the corresponding comments always vanished after
expiration.  Furthermore, the expiration of one acknowledgement didn't
remove all other acknowledgements for the same host/service.

Both aspects were inconsistent with the behavior of the manual removal of
acknowledgements, which was quite confusing and unintended.  This has been
fixed, so acknowledgement expiry is identical to manual removal at the given
point in time.

Change-Id: If5db0fdeb2268988fbf68fccdc8a7f547d5015fe