11938 REST API: make optimistic locking via ETag configurable
When multiple HTTP clients want to update an object at the same time,
it can happen that the slower client will overwrite changes by the faster one.
This is commonly referred to as the 'lost update problem'. To prevent this
situation from happening, Checkmk's REST API does 'optimistic locking' using
HTTP ETag headers. In this case the Object's ETag has to be sent to the server
with a HTTP If-Match header.

This behavior can now be deactivated, but this will allow the 'lost update
problem' to occur.

Change-Id: I3cdd52c349cba8d0f532addd921af89639cce640