12010 FIX Fix for config update: Regex validation will ignore items that are not dictionaries
Since checkmk version 2.0.0b7 the Checkmk update procedure (cmk-update-config) validates regexes.
This validation failed with the error message "AttributeError: 'str' object has no attribute 'get'"
if the item of a rule was not of the type "dict" or "None".
This werk will skip the regex validation if the item is not of type "dict".

Change-Id: I1597359fea771cb19085e92171dd18d4b90edbbd