11978 ceph_df: support for octopus and nautilus with JSON output
This werk will make the ceph plugin more resilient against subtle
output changes in future versions of the ceph command line tool.
When the new agent plugin is deployed the numbers that are reported by the ceph checks
will change slightly, because the json output is more precise.
If the JSON section cannot be generated the old output method without json will be used.

Change-Id: Idf358d68dc790d5cb582ba16d9e7a0cd1592d31b