10614 FIX Improved hexagon visualization on dashboards
Host/Service statistics dashlet:
LI: Up hosts and OK services are now displayed in green again, not in grey as was recently implemented.

Site/Host overview dashlet:
LI: The dashlet is now responsive to dashboard filters.
LI: Host and service state colors are now in line with Checkmk standards. Before, down hosts were shown in yellow (now red) and unknown services as well as unreachable hosts in red (now orange).
LI: The positioning of hexagons in the dashlet was fixed so they do not overflow beyond the dashlet borders anymore.

Change-Id: I3ea04ff13e973fb5a05eb11fa38b9e4f81f2ab57