Fix layout of dataset layout for tables
Tables that used the dataset layout (like e.g. the host status view)
used all available space for the first column which only contains
the title of each row. As a result the corresponding value for the
row was shown at the far right. Now the title column has the width
of the longest title.

Change-Id: I240876384b6ca4ef10404e449329832a3c19a94f