11791 FIX First column in single dataset layout of views consumed too much space
This does only affect the layout mode "Single dataset". A view that uses this
layout is e.g. "Status of host HOSTNAME". The first column that displays the
title of the row consumed too much space. In consequene the value
corresponding to that title was pushed to the right side of the window.

Change-Id: I6870c483a1339e92884c6dceb60cfc72778c0f36