Title headers in views can now have extra CSS classes
This commit adds the method title_classes to Painter
and its subclasses. This method is then used to
align the text of table headers for some painters on
the right. Previously, the headers were aligned left
whereas the data of the column was aligned right.
An example for this is the "All hosts" view.

Change-Id: Iabc93d85c687660d90615c192e788178f0501854