CMK-7227: Add agent rule for setting python version
With the help of the "python_plugins" agent rule/bakery plugin,
it is possible to choose the python version, under that python
plugins will be executed on the host systems, rather than letting
the agent detect it automatically.
Furthermore, it is possible to optionally set an alternative python
command, that will be used for the execution.

Change-Id: I0f206057e998158e6e7e0c58bc4b96c8a6985f74