CMK-7274: UNIX-Agent: Use the determined python command for plugin execution
Execute python plugins with the help of the determined python
command (i.e., "python3", "python2" or "python") instead of executing
it directly as a script.

This fixes situations in which a Python 2 installation is found via
the "python" command, but the plugin is executed with "python2" command
taken from the shebang.

Change-Id: I61cbdca2ad6512504ad37ff8556647dc63f3713f