CMK-7274: UNIX-Agent: Use the determined python command for plugin execution
Execute python plugins with the help of the determined python
command (i.e., "python3", "python2" or "python") instead of executing
it directly as a script.

This fixes situations in which a Python 2 installation is found via
the "python" command, but the plugin is executed with "python2" command
taken from the shebang.

Change-Id: Ida35e932c96ab08c7f7421106fe9381b189ad63c