12143 Graph dashlets: Optionally disable time range synchronization
When a dashboard has multiple graph dashlets showing different time ranges, it
may be not intended to synchronize the time ranges by zooming into a single
graph. By default Checkmk synchronizes the time ranges and pins of all graphs
on a page.

This can now be disabled for each graph dashlet individually by setting the
graph rendering option "Do not follow timerange changes of other graphs on the
current page".

Plus: Set the new option for the graph dashlets on the new main
dashboard.

Change-Id: I3e3e0663f9526d8b9c02d1902c78a957b2c73e8e