12177 FIX Process management board data in HW/SW inventory
If management boards where not configured as a separate host, but as
an additional data source for a regular host, the data provided by them
has not been processed by the hardware/software inventory in checkmk
versions 2.0.0i1 to 2.0.0b8.

This fix is not as blunt as it may seem: The behaviour implemented by
this extra `for` loop is as close to the checking behaviour as we can
get in the inventory world.

Change-Id: Ifc9bc92a52b66a95dcee396fb5ae02a067b1948d