Default column type/namespace to non-templatized version
 - DoubleColumn<T>::Constant -> DoubleColumn::Constant
 - DoubleColumn<T>::Reference -> DoubleColumn::Reference
 - DoubleColumn<T> -> DoubleColumn::Callback<T>

Removing the template parameter from the base name is more flexible.
Moreover, it is true that the Constant and Reference cases do not
need T as they do not call `columnData<>`.

Apart from the changes in `DoubleColumn.`, the patch set is mechanical
and was performed with PCRE on the repo.

CMK-7076

Change-Id: Ia92fec0e2015bc50d817f1ea48e559a6041bba0a