12181 FreeBSD interfaces: prefer SNMP check
This change only affects hosts that are monitored using
SNMP data <b>and</b> data provided by the Checkmk agent
for FreeBSD.

In these cases we now prefer the services discovered by
the {{if64}} check over the ones discovered by the
{{interfaces}} check using the {{statgrab_net}} section.

The same rulesets and parameters apply.

If this affects you, we recommend you rediscover the new
services, as the SNMP services provide more information
in this case.

However, if you want to continue using the services
driven by the {{statgrab_net}} section, you can disable
the SNMP section {{if64}} for the affected host using
the rule "Disabled and enabled sections (SNMP)".

Change-Id: I30a605720bde382384543664d0c037c64d4a225e