Let List{Lambda,}Column inherit Column directly
We remove the coupling in order to have more freedom in
List{Lambda,}Column to remove the original ListColumn.

There, we move the original ListColumn into the `deprecated`
namespace and follow the actual naming conventions for the
lambda and constant columns.

 - ListColumn -> deprecated::ListColumn
 - ListLambdaColumn -> ListColumn::Callback

Add ListColumn::Constant and ListColumn::Reference and tests for
consistency.

CMK-7076

Change-Id: I81c3cefe6f5c14b80dca8cb143f5145f2464358e