Log MKFetcherError at INFO
Our MKFetcherError exceptions are used to differentiate between
bonehead exceptions and errors that are exceptional but not
unexpected, such as I/O error for TCP, &c.

These MKFetcherError are then special cased in the summarizers
and lead to precise messages.  However, such errors are typically
not critical to the monitoring.

We therefore set the logging level for such exceptions to INFO
(was CRITICAL) in order to avoid spamming the cmc.log.

FEED-5520

Change-Id: I89a2d6a603a71dcb009f2f8baaace5e34ea56bcd