12148 FIX Bake&Sign sometimes misses some signatures
In some situations, the agent bakery will fail to sign some agent
packages after baking in "Bake and sign" mode.

This may occur when a host is added to an already existing agent
configuration, resulting in a rebaking of the package under the
same configuration hash.

With this Werk, all agents will be signed when baking in
"Bake and sign" mode.

Change-Id: I037fbf8371453bf701c911d1a781c6f214223d14