12183 FIX Outdated data during full scan discovery
During full scan discovery on the WATO discovery page
outdated data has been used. While this was not
noticable in most cases, it could lead to missing
or unexpectedly present services.
This was the case for instance if you disabled SNMP
sections, added a new plugin or changed an SNMP
detection specification.

Change-Id: I8b1adf6c62356563c325f9b8942e6cd6079c3e53