12281 FIX Activate changes: Failed to report "locked site" state
The activate changes site failed to report the state of the activation when
another activation is currently running on a site while another activation was
started.

In this case the activation background job failed with an error like "raise
Exception("start time not set")". Instead of this message the error is now
shown to the user as activation details message.

Change-Id: I1ea3675c5545fd06caefe26c0b4afc58f16a95ac