Main dashboard is only permitted for admins and guests
Special hack for the "main" dashboard: It contains graphs that are only
correct in case you are permitted on all hosts and services. All
elements on the dashboard are filtered by the individual user
permissions. Only the problem graphs are not able to respect these
permissions. To not confuse the users we make the "main" dashboard in
the enterprise editions only visible to the roles that have the
"general.see_all" permission

Change-Id: Ib5033e0aa254ee2ef269afbdd5d81ad42063eb92