12273 FIX API: modifications for downtime endpoint
This werk introduces a few changes to the downtime endpoint. Proper status codes
are now returned when trying to set a downtime on non existing host and service groups.
A 422 status code is returned when the provided query for setting a service or host
downtime returned an empty list. The type 'by_id' is now an enabled option for
the delete endpoint. The parameters 'hostname' and 'services' in the delete endpoint
were renamed to 'host_name' and 'service_descriptions' to introduce more consistency.
Previously, using the delete endpoint with a host_name via the delete_type 'params' resulted
in the deletion of all downtimes that contained the specified host_name, regardless of
service or host. This has now been changed to be restricted to host downtimes only.

Change-Id: I2b92baa6d89b11923f395170fa59f6cc9c55a478