11853 FIX Windows agent uses separate script to protect own files from unathorized access
Previously, the protection algorithm was hard-coded into the Windows agent and
it was called synchronously. Such approach may lead to some rare but important
problems, because the start of the Windows agent service required a lot of
time and Windows service manager may not accept this delay.

Since this release Windows agent uses batch script, which runs in a separate
and independent process. This script provides the identical functionality as
before, but it doesn't interfere with Windows agent service thus solving the
problem with delay.

Change-Id: Idc604763b5d3c04ed05fa873f37566257b16d5f8