Fix height of input, select and texarea fields
The size of elements may be specified in valuespecs and is then set
via the style attribute of the element in the Python code. In a
previous commit the size of elements was changed to a border-box.
Accoring to the CSS Box Sizing Module Level 3:

The content width and height are calculated by subtracting the
border and padding widths of the respective sides from the specified
<length-percentage>. As the content width and height cannot be
negative, this computation is floored at zero.

I.e. content that did not fit in the border-box was truncated.
This commit restores the old width of elements while it keeps the
new height intact.

Change-Id: I69636e5081ec9e6bb0c4ca0a18efb2bba2de9682