drop no-op statement
This function is supposed to return the present host labels.
Due tue a refactoring, it seemed that it returned the
vanished host labels as well, but this was a red herring:
We do not load the present host labels in this function.
If the set of 'old' host labels is empty, we cannot have
any 'vanished' host labels.

Change-Id: I748c1b7150bb35e00ccc1ed5fad08df940f9d1ba