11856 FIX Windows agent installation now supresses reboot of a host even if reboot is requested
Windows msiexec may request a reboot if some of the Windows agent's files(directories) were locked during installation procedure.
Such reboot is performed automatically and immediately when the installation is silent. And silent installation is a typical use
case for Windows agent update routine.

Evidently, this kind of spontaneous host's reboot is inacceptable in most cases and since this release will be fully suppressed.

It is important to note that although a reboot will be prevented, it is the user's responsibility to reboot the host at
convenient time. It is also strongly recommended to find out which software and for what reason locks the files/directories in
the Windows agent's directories.

Change-Id: I29fbb9486899bebdc3abbe10245c46c37844dc2b