Force perfomance graphs t-axis always with labels
The way ticks & labels are calculated in our performance graphs does not
enforce the option of actually labeling the edges(start & end ticks). Yet
setting a number of labels of 2 promotes that half of the range could be at
least be covered by a label, somewhere in the timeline. Problem of division
by zero on graph is solved, yet on very narrow graphs in dashlets it can
still happen that no label is drawn.

Performance graphs will be replaced by D3js graphs, which have also solved
the problem of rendering human time. From seconds to months and years which
have different length. This fix is only to fix the case where no labels are
expected and thus the graph changes height.

Change-Id: I9089be020e25872d7bf1700d0c85b283b53f4de4