12337 FIX REST API acknowledgment status codes
When acknowledging a non existing service, a HTTP 404 was about
to be triggered, but was not defined for this endpoint. This could
lead to an HTTP 500 (Internal Server Error). This has now been
fixed to correctly return a HTTP 400 (Bad Request) error, which
is defined for this endpoint, brining it in line with the documentation.

Change-Id: I80bd551e6599469ee4d665c8ea829877d045985b