12294 FIX Changing predefined conditions now triggers core reload
Changing the conditions of a predefined condition did not trigger a core reload
during the following activate changes. This made the change appear not not
affect existing rules. In fact it was just delayed until the next core config
creation + reload. This has now been changed to directly triggering a core
config creation + reload.

Change-Id: I694569a468a1e1ac8b876c509a2d3345d6f99eb6