Label discovery: Fix compatibility with 2.0.0 central site
When a site as of 2.0.0p1 is talking with an older 2.0.0 central site,
it must not send it the new data structures that only the 2.0.0 is aware
of.

To make it possible to detect the old site we now use the HTTP header
"x-checkmk-version" which is sent with all automation calls as of
version 2.0.0p1. When such a header is missing the remote site knows it
is talking to an older site and must not send the new data structures.

Change-Id: I9fbad97d71ebbf60456e8cd6ddd57a73f65064ea