12375 API host_config: modifying the folder link of the served host
In the previous version, the folder link of the served host was part of
the members. This is now moved to the links with this change. In addition,
the original rel is changed from using the folder id to using the canonical
path of the folder. Since the canonical path of the folder may change during
the course, an additional link (permalink) is add based on the folder id. Both
links can be used to retrieve the folder config.

Change-Id: Ia639e8445a6466f08b5f85781e3928aaceeb0681