Improve configuration lock error message
In case the configuration lock is being acquired by a process, e.g.
during activate changes while a longer running other process, like
baking of agents, is running, an error message was shown to the user
which was a bit cryptic and also offered the option to report this
issue as a crash report.

We now display an error message which should be easier to understand:

Couldn't lock the Checkmk configuration. Another process is running that
holds this lock. In order for you to be able to perform the desired
action, you have to wait until the other process has finished. Please
try again later.

Change-Id: I3c2af5b97d677c6d406992956564fbbe015dce4b