12552 FIX Notes URL is now always opened in new tab or window
You can configure a "Custom notes URL" for host and services.  A link to this
notes URL appears in the host or service context menu in the GUI.

With the introduction of the Content Security Policy (which is configured in
etc/apache/conf.d/security.conf), it is now not possible anymore to follow
external links within the GUI frameset without changing the policy to allow
links to the linked domain.

In such a case the browser console shows something like "Content Security
Policy: The page’s settings blocked the loading of a resource ...".

To make the "Custom notes" link work by default we now changed that link to
open the link target in a new tab.

Change-Id: Id19a0431afc4aef34150b08df3e8f6f4779617b1