12421 FIX Notification history: links for multiple contacts
The views showing the notification histories of hosts and/or services
("Notifications of host", "Notifications of host & services") display the
contacts to which the notifications were sent. These contacts are linked
to the view showing only the notifications sent to these contacts. However,
in case a notification was sent to more than one contact, the individual
contacts listed in the notification history were not clickable individually,
as one would expect. Instead, only the entire list of all contacts was
clickable and the target page showed the notifications which were sent to all
of the contacts. Now, each contact is clickable on its own and the target page
shows only the notifications which were sent to this particular contact.

SUP-3631

Change-Id: I4f712fe070bcd8185bef7ad44c2ce1a33d0ec08a