RRDtool cli xport patch
This patch was not merged
https://github.com/oetiker/rrdtool-1.x/pull/980#issuecomment-474408355

The rrdtool xport tool on the cli does not present data as it is stored in
memory, which is what CMK does receive on a livestatus query.

For debugging purposes it is best to keep these two outputs in sync, even
if RRDTool does not display them identically.

Integration test checks that inmemory, python ext and cli return the same
metadata and data.

Change-Id: If5acb18220ae6a683ab21fdeb94d6cf41fbc7a47