12420 FIX Dashboards: Unify handling of dashboard-wide filters
There are two ways to define dashboard-wide filters:
<ol>
  <li>In the dashboard configuration, these filters are persisted.</li>
  <li>Using the filter menu, these filters are not persisted.</li>
</ol>

This werk allows users to consistenly overwrite filters defined via option 1
with filters defined via option 2. Previously, this was only possible for some
dashlets. Note that this overwrite is of course only temporary. Once the
page showing the dashboard is closed, any filters defined via option 2 are
gone.

Note that filters defined on the dashlet level still have the highest priority,
as was the case before this werk. This means that any filter defined via option 1
or 2 from above will not overwrite a filter defined in the configuration of an
individual dashlet.

FEED-5488

Change-Id: I69aef3742c9ae49a2be9703a7d2432832d2a6239