12423 <tt>cisco_asa_failover</tt>: Also monitor local, remote and failover link status
This werk extends the functionality of the check plugin <tt>cisco_asa_failover</tt>,
which monitors the status of Cisco ASA failover clusters. This plugin will now go
WARN if the local or the remote device are not either active or in standby.
Furthermore, the the plugin will now report CRIT if the failover link is not up.

This change also migrates cisco_asa_failover to the new API.

Closes #350

Change-Id: I1be02ff95356029281d8947028c260eb6fbbc22a