12553 FIX Agent bakery: Fix host label conditions in distributed setups
In previous Checkmk versions agent bakery rules that were using host label
conditions could not be used correctly in distributed setups.

The background is that in distributed setups the discovered(!) host labels are
stored on remote sites for hosts that are assigned with a remote site.  But the
processing of agent bakery rules is done one the central site during agent
baking. Without the discovered host labels of remote site hosts, the bakery
could not compute the correct agent configuration for these hosts. The hosts
were not matched based on their discovered host labels.

This change now sets up a synchronization of discovered host labels from remote
sites to the central site for the bakery rules to be able to match the correct
hosts.

The synchronization is being executed on a regular base (every 10 minutes for
now) as background job. You can have a look at the state of this
synchronization at "Setup > Background jobs > Discovered host label
synchronization".

Change-Id: Id7112edf0f91a5d2c46d5bfd566defea0ec6b645