12309 FIX nagios core: include legacy dependencies of new api plugins
This only applies to the raw edition with nagios core: If a new style
check plugin is interested in multiple sections, and one section is a
legacy style check, this section was not parsed at all and the string
table of the section was forwarded to the new style api check, which then
crashes as it expected the parsed data.

For example: The diskstat check (new style api) is dependent on multipath
(legacy check) and crashed if the multipath check was disabled.

This should now be fixed.

CMK-7489

Change-Id: Ie1a5b6842a15b0ae54b83837bc180c2dae563357