12643 FIX oracle_asm_diskgroup: fix crash "TypeError (unsupported operand type(s) for *: 'NoneType' and 'int')".
If the checkbox for the parameter "Trend computation" of the rule "ASM Disk Group (used space and growth)" was enabled,
but the parameter "Enable trend computation" (trendrange) was left unchecked the check oracle_asm_diskgroup
crashed with the error message: "TypeError (unsupported operand type(s) for *: 'NoneType' and 'int')".
This condition is fixed by falling back to default parameters.

Change-Id: I32a446d0f71db97f12ce23525ff9bdea57ae097a