Define more type aliases for {Double,Int}Column
We define `Default`, `function_type`, and `column_type`.  All three are
consistent with similar names in other columns, such as `ListColumn` or
`TimeColumn`.

The added consistency makes the code easier to understand in that some
functionality in the classes now uses the exact same code:  understand
one---know them all!

This is also a slightly higher level of abstraction in that the names
reflect the functionality/purpose instead of the concrete type.

CMK-7076

Change-Id: I2c5f964fbbc4c56eb9904d46b071bd804ff035b8