12554 FIX Ruleset editor: Fix analyzing host label matching in distributed setups
The ruleset editor provides a mode that allows a user to analyze the matching
of rules on specific hosts services. This mode analyzes the conditions configured
for a rule and reports whether or not these conditions match on the given host.

One of these conditions can be host labels. In distributed setups the host labels
are stored on the site the host is monitored from, while the analyzation of the
condition matching is being done on the central site.

If you have rules that have host labels configured as conditions, analyzing
the rule matching did not work because of the missing discovered host labels in
the central site.

This is now solved by synchronizing the host labels from the remote site to the
central site in the moment they are needed. This mechanism is based on the
synchronization mechanism that was introduced with werk #12553.

Change-Id: Icc7323c159bbec948c24c8d3672922723d1617d2